requirements_local.txt
fastapi
uvicorn
numpy
pillow
python-multipart
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.0.0+cpu
torchvision==0.15.0+cpu